UI Updated for RentalListingDetailsView

+++ SecureRental/Models/User.swift
@@ -6,7 +6,7 @@ class AppUser: ObservableObject, Identifiable, Codable {
     var email: String
     var name: String
     var profilePictureURL: String?
-    var rating: Int               // 1 to 5
+    var rating: Double               // 1 to 5
     var reviews: [String]         // Multiple reviews
     var favoriteListingIDs: [String] = [] 
     
@@ -17,7 +17,7 @@ class AppUser: ObservableObject, Identifiable, Codable {
         email: String,
         name: String,
         profilePictureURL: String? = nil,
-        rating: Int = 0,
+        rating: Double = 0.0,
         reviews: [String] = [],
         favoriteListingIDs: [String] = []
     ) {
